+++ tests/node_parser.test.ts
@@ -2,8 +2,8 @@ import {
   StringASTNodeImpl,
   PropertyASTNodeImpl,
   ObjectASTNodeImpl,
-} from './ast_node_impl';
-import NodeParser from './node_parser';
+} from '../src/ast_node_impl';
+import NodeParser from '../src/node_parser';
 
 describe('NodeParser', () => {
   // String node
